requirements.txt
CairoSVG==2.5.1
env-canada==0.0.35
ephem==3.7.7.0
fastjsonschema==2.14.4
geocoder==1.38.1
gpiozero==1.5.1
noaa-sdk==0.1.18
pillow>=6.2.2
printtools==1.2
PyInstaller==3.6
pyowm>=3.0.0
python-tsl2591==0.2.0
questionary==1.5.2
regex==2020.4.4
RPi.GPIO==0.7.0
APScheduler==3.6.3
lastversion==1.1.6
